Revert "RWLock: don't assign the lockdep id more than once"
This reverts commit f6cc9789a3f69f8411c6a80cc496d204e68ab58c.

It's important to assign these for all operations for cases where
g_lockdep isn't yet true when the constructor runs.  This is true
for the HeartbeatMap rwlock, among other things, as that thread
is created during early startup before lockdep is enabled.  All
of the lockdep hooks assume that they can assign ids on the fly
and not tracking them here breaks things.

Conflicts:

	src/common/RWLock.h