osd: loop over uint8_t instead of shard_id_t
When looping in the range [0,number of shards[ using a uint8_t variable
which is an integer instead of shard_id_t. The alternative would be
define the ++ operator on shard_id_t but since the two are used
interchangeably in a number of places, that seems unnecessary.

Signed-off-by: Loic Dachary <loic@dachary.org>
(cherry picked from commit 20e66d16094b1ef5e1038485f264a446016201ea)