krbd: catch up with run_xfstests.sh changes
In the current xfstests, tests cannot be listed in ranges.  Because of
this, and because our distribution of tests between our three clients
was pretty uneven, I've changed this job to run the whole auto group
(with some tests excluded by run_xfstests_krbd.sh) in random order on
two clients in parallel.  It takes longer to run, but it also spares
a test box and can potentially create a better test workload.

Signed-off-by: Ilya Dryomov <ilya.dryomov@inktank.com>