qa: cp run_xfstests.sh run_xfstests-obsolete.sh
run_xfstests.sh is going to be updated in the next commit to be able to
drive newer xfstests.  Among other things, the new xfstests proper
doesn't support listing tests in ranges, which is what the qemu wrapper
(run_xfstests_qemu.sh) relies on.  So keep a copy of the old
run_xfstests.sh around until the qemu vm image is regenerated and the
up-to-date exclusion list for that kernel is shaken out.

Signed-off-by: Ilya Dryomov <ilya.dryomov@inktank.com>