Added a note on keyring location
This quick start says that ceph-deploy will copy over the keyring. It neglects to mention the location. This makes it really confusing when 'Configure a Block Device' uses a dummy keyring directory: '/path/to/keyring'.

The note added to the end of the 'Install Ceph' section will hopefully clear this up.