osd: workaround race condition in tests
Trying to "ceph tell" a newly created OSD sometime triggers an
ENXIO. The OSD creation function used for test scripts waits for the OSD
to report that it is up (according to ceph osd tree) before
returning. That reduces (maybe close ?) the condition that triggers the
error so that the tests do not randomly fail.

http://tracker.ceph.com/issues/8630 refs: #8630

Signed-off-by: Loic Dachary <loic@dachary.org>